feat: add useStudentDetail hook
